serialize bgs response
