+++ modules/veteran_verification/app/controllers/veteran_verification/v0/disability_rating_controller.rb
@@ -8,33 +8,22 @@ module VeteranVerification
     class DisabilityRatingController < ApplicationController
       include ActionController::MimeResponds
 
-      NOTARY = VeteranVerification::Notary.new(Settings.vet_verification.key_path)
-
-      before_action { authorize :evss, :access? }
-      before_action { permit_scopes %w[disability_rating.read] }
-
       def index
-        disabilities_response = service.get_rated_disabilities
+        disabilities_response = rating_service.get_rating(@current_user)
         serialized = ActiveModelSerializers::SerializableResource.new(
-          disabilities_response.rated_disabilities,
-          serializer: VeteranVerification::DisabilityRatingSerializer
+            disabilities_response,
+            each_serializer: VeteranVerification::DisabilityRatingSerializer
         )
 
         respond_to do |format|
           format.json { render json: serialized.to_json }
-          format.jwt { render body: NOTARY.sign(serialized.serializable_hash) }
         end
       end
 
       private
 
-      def service
-        EVSS::DisabilityCompensationForm::Service.new(auth_headers)
-      end
-
-      def auth_headers
-        headers = EVSS::DisabilityCompensationAuthHeaders.new(@current_user)
-        headers.add_headers(EVSS::AuthHeaders.new(@current_user).to_h)
+      def rating_service
+        BGS::DisabilityRatingService.new
       end
     end
   end
