+++ modules/veteran_verification/app/serializers/veteran_verification/disability_rating_serializer.rb
@@ -1,12 +1,31 @@
 # frozen_string_literal: true
 
+
 module VeteranVerification
   class DisabilityRatingSerializer < ActiveModel::Serializer
-    attributes :rating_percentage, :effective_date, :decision
-    type 'disability_ratings'
+    include ActiveModel::Serialization
 
-    def decision
-      object.decision_text
+    type :disability_ratings
+
+    attribute :overall_disability_rating
+    attribute :ratings
+
+    def id
+      nil
+    end
+
+    def overall_disability_rating
+      object[:disability_rating_record][:service_connected_combined_degree]
+    end
+
+    def ratings
+      object[:disability_rating_record][:ratings].map do |rating|
+        Rating.new(
+          decision: rating[:disability_decision_type_name],
+          effective_date: rating[:begin_date],
+          rating_percentage: rating[:diagnostic_percent]
+        )
+      end
     end
   end
 end
